took Home component out of index.js, made Home.js, with http calls
